+++ pyflysight/__init__.py
@@ -1,0 +1,6 @@
+from enum import IntEnum
+
+
+class FlysightType(IntEnum):  # noqa: D101
+    VERSION_1 = 1
+    VERSION_2 = 2