+++ CHANGELOG.md
@@ -1,2 +1,11 @@
 # Changelog
 Versions follow [Semantic Versioning](https://semver.org/spec/v2.0.0.html) (`<major>`.`<minor>`.`<patch>`)
+
+## [v1.0.0]
+### Added
+* Add parsing pipelines for Flysight V1 track data files
+* #8 Add parsing pipelines for Flysight V2 track & sensor data files
+* #11 Add trinning for Flysight V1 & V2 data files
+
+## [v0.1.0]
+Initial release - yay!